requirements.txt
streamlit>=1.28.0
langchain>=0.0.340
langchain-openai>=0.0.2
python-dotenv>=1.0.0
tiktoken>=0.5.1
openai>=1.3.0
pydantic>=2.0.0